Use x/net/context and prepare code for gRPC
This commit makes changes that prepare the codebase for gRPC.

The standard library context is not yet used in the gRPC package so it
forces us to use golang.org/x/net/context in the whole codebase to
simplify things.

At the same time, when defining the services in protobuf format, we need
to be explicit about our types thus limit and offset are changed to
int64 to avoid conversions between the generated gRPC code and our
normal code.

By choosing to use gRPC we automatically introduce a lot of dependencies
in the project which are required by the gRPC package. On top of that,
the complexity is increased in comparison to REST/JSON as additional
steps are introduced, including generating stubs and gateways. The main
benefit is that there is a main schema in protobuf format from which
both gRPC and REST APIs are generated along with the rest of the
benefits that gRPC brings.